requirements.txt
grad_cam==1.5.5
pandas==2.3.2
Pillow==11.3.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
